Ensure files in the archive are user-writable.
Squirrel.Mac tries to remove the quarantine xattr on all files when it
applies the update; this fails if any file in the update is not user-
writable.  This ensures that we package all files (for eventual
distribution in Rancher Desktop) as user-writable, so that it can
successfully remove the extended attribute.

Signed-off-by: Mark Yen <mark.yen@suse.com>